+++ cmp/grid/Grid.ts
@@ -8,9 +8,8 @@ import {GridApi, AgColumnState} from '@xh/hoist/kit/ag-grid';
 
 import {agGrid, AgGrid} from '@xh/hoist/cmp/ag-grid';
 import {ColumnState, getTreeStyleClasses} from '@xh/hoist/cmp/grid';
-import {gridHScrollbar} from '@xh/hoist/cmp/grid/impl/GridHScrollbar';
 import {getAgGridMenuItems} from '@xh/hoist/cmp/grid/impl/MenuSupport';
-import {div, fragment, frame, vframe} from '@xh/hoist/cmp/layout';
+import {div, fragment, frame} from '@xh/hoist/cmp/layout';
 import {
     hoistCmp,
     HoistModel,
@@ -98,19 +97,11 @@ export const [Grid, grid] = hoistCmp.withFactory<GridProps>({
     className: 'xh-grid',
 
     render({model, className, testId, ...props}, ref) {
-        const {
-                store,
-                treeMode,
-                treeStyle,
-                highlightRowOnClick,
-                colChooserModel,
-                filterModel,
-                enableFullWidthScroll
-            } = model,
+        const {store, treeMode, treeStyle, highlightRowOnClick, colChooserModel, filterModel} =
+                model,
             impl = useLocalModel(GridLocalModel),
             platformColChooser = XH.isMobileApp ? mobileColChooser : desktopColChooser,
-            maxDepth = impl.isHierarchical ? store.maxDepth : null,
-            container = enableFullWidthScroll ? vframe : frame;
+            maxDepth = impl.isHierarchical ? store.maxDepth : null;
 
         className = classNames(
             className,
@@ -122,17 +113,13 @@ export const [Grid, grid] = hoistCmp.withFactory<GridProps>({
         );
 
         return fragment(
-            container({
+            frame({
                 className,
                 items: [
                     agGrid({
                         model: model.agGridModel,
                         ...getLayoutProps(props),
                         ...impl.agOptions
-                    }),
-                    gridHScrollbar({
-                        omit: !enableFullWidthScroll,
-                        gridLocalModel: impl
                     })
                 ],
                 testId,
@@ -312,11 +299,6 @@ export class GridLocalModel extends HoistModel {
                 treeData: true,
                 getDataPath: this.getDataPath
             };
-        }
-
-        // Support for FullWidthScroll
-        if (model.enableFullWidthScroll) {
-            ret.suppressHorizontalScroll = true;
         }
 
         return ret;